+++ src/main/java/com/mychess/my_chess_backend/utils/enums/ChessPieceColor.java
@@ -3,7 +3,7 @@ package com.mychess.my_chess_backend.utils.enums;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 public enum ChessPieceColor {
-    WHITE("c"),
+    WHITE("w"),
     BLACK("b");
 
     private final String value;
